chore: refactor account priority in session
